# api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: User Service
  description: serve user authentication API
servers:
  - url: http://localhost
paths:
  /register:
    post:
      operationId: Register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserData'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'

  /login:
    post:
      operationId: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginData'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'

  /user:
    get:
      operationId: GetUserData
      security:
        - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
    post:
      operationId: UpdateUserData
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserData'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'

components:
  schemas:
    NewUserData:
      type: object
      required:
        - name
        - phone
        - password
      properties:
        password:
          type: string
        name:
          type: string
        phone:
          type: string

    UserData:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string

    LoginData:
      type: object
      required:
        - phone
        - password
      properties:
        phone:
          type: string
        password:
          type: string 

    LoginResponse:
      type: object
      required:
        - token
        - message
      properties:
        token:
          type: string
        message:
          type: string
  
    UserResponse:
      type: object
      required:
        - name
        - phone
        - message
      properties:
        name:
          type: string
        phone:
          type: string
        message:
          type: string

    Response:
      type: object
      required:
        - message
      properties:
        message:
          type: string

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer